# Discard all bytes read in the handler benchmark.