<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

The httptest.ResponseRecorder reads the entire response in-memory, which is fine in general for tests but greatly dilutes the results of a benchmark. This forwards all bytes to a Discard writer to be able to accurately show what the handler will allocate in production scenarios.

Output for posterity:

### Without BufferPool

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	   97948	     12211 ns/op	   35926 B/op	      36 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	  201487	      5172 ns/op	   35765 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  108157	     12431 ns/op	   35860 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	  228391	      5275 ns/op	   35901 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	   98913	     13133 ns/op	   35919 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	  221556	      5306 ns/op	   35922 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	   82387	     15294 ns/op	   35876 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	  221085	      5755 ns/op	   35785 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	   61533	     18599 ns/op	   35901 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  192728	      6300 ns/op	   35928 B/op	      36 allocs/op
PASS
```

### With BufferPool

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	  187291	      6285 ns/op	    3099 B/op	      36 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	  915727	      2150 ns/op	    3128 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  178771	      6670 ns/op	    3126 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	  910431	      1865 ns/op	    3132 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	  167673	      7292 ns/op	    3167 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	  836035	      1974 ns/op	    3189 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	  131922	      8476 ns/op	    3169 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	  688290	      2001 ns/op	    3170 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	  108840	     10673 ns/op	    3137 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  597840	      2476 ns/op	    3117 B/op	      36 allocs/op
PASS
```

This now very clearly shows the allocation behavior differences between a version that runs with the BufferPool and one without it.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov